+++ packages/pioneer/pioneer-react/src/index.tsx
@@ -522,7 +522,6 @@ export const PioneerProvider = ({ children }: { children: React.ReactNode }): JS
       //TODO wallet or wallet type? cache per device? or wallet type?
       const cacheKey = `cache:blockchains:${walletType}`;
       const cachedBlockchains: string[] = JSON.parse(localStorage.getItem(cacheKey) || '[]');
-
       console.log(tag, 'cachedBlockchains: ', cachedBlockchains);
 
       const getNetworkIdFromChainStr = (chainStr: string): string | undefined => {
@@ -566,6 +565,36 @@ export const PioneerProvider = ({ children }: { children: React.ReactNode }): JS
       // Fetch paths based on blockchains
       const paths: any = getPaths(blockchains);
       console.log(tag, 'paths: ', paths);
+
+      //add custom paths
+      paths.push({
+        note: 'Bitcoin account 0 segwit (p2sh)',
+        networks: ['bip122:000000000019d6689c085ae165831e93'],
+        script_type: 'p2sh-p2wpkh',
+        available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        type: 'xpub',
+        addressNList: [0x80000000 + 49, 0x80000000 + 0, 0x80000000 + 0],
+        addressNListMaster: [0x80000000 + 49, 0x80000000 + 0, 0x80000000 + 0, 0, 0],
+        curve: 'secp256k1',
+        showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+      });
+
+      paths.push({
+        note: 'Bitcoin account 1 Native Segwit (Bech32)',
+        blockchain: 'bitcoin',
+        symbol: 'BTC',
+        symbolSwapKit: 'BTC',
+        network: 'bip122:000000000019d6689c085ae165831e93',
+        script_type: 'p2wpkh', //bech32
+        available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        type: 'zpub',
+        addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
+        addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
+        curve: 'secp256k1',
+        showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+      });
+
+      //TODO get user added paths
 
       const spec =
         localStorage.getItem('pioneerUrl') ||