+++ packages/wallets/okx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @coinmasters/wallet-okx
 
+## 11.0.37
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@11.0.37
+  - @coinmasters/toolbox-utxo@11.0.37
+  - @coinmasters/types@4.7.37
+  - @coinmasters/toolbox-evm@11.0.37
+
 ## 11.0.36
 
 ### Patch Changes